+++ mglogger/src/main/java/com/mgtv/logger/java/OnLoganProtocolStatus.java
@@ -20,21 +20,9 @@
  * THE SOFTWARE.
  */
 
-package com.mgtv.logger;
+package com.mgtv.logger.java;
 
-public interface LoganProtocolHandler {
-
-    void logan_flush();
-
-    void logan_write(int flag, String log, long local_time, String thread_name,
-            long thread_id, boolean is_main);
-
-    void logan_open(String file_name);
-
-    void logan_init(String cache_path, String dir_path, int max_file, String encrypt_key_16,
-            String encrypt_iv_16);
-
-    void logan_debug(boolean debug);
-
-    void setOnLoganProtocolStatus(OnLoganProtocolStatus listener);
+public interface OnLoganProtocolStatus {
+    void loganProtocolStatus(String cmd, int code);
 }
+
