+++ mglogger/src/main/java/com/mgtv/logger/java/SendLogDefaultRunnable.java
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.mgtv.logger;
+package com.mgtv.logger.java;
 
 import android.util.Log;
